@@ -1,12 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
+import 'package:localization_riverpod_flutter/src/counter_page/counter_page.dart';
 import 'package:localization_riverpod_flutter/src/localization/app_localizations_context.dart';
 
-import 'sample_feature/sample_item_details_view.dart';
-import 'sample_feature/sample_item_list_view.dart';
-
 class MyApp extends StatelessWidget {
-  const MyApp({Key? key}) : super(key: key);
+  const MyApp({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -15,20 +13,7 @@ class MyApp extends StatelessWidget {
       localizationsDelegates: AppLocalizations.localizationsDelegates,
       supportedLocales: AppLocalizations.supportedLocales,
       onGenerateTitle: (BuildContext context) => context.loc.appTitle,
-      onGenerateRoute: (RouteSettings routeSettings) {
-        return MaterialPageRoute<void>(
-          settings: routeSettings,
-          builder: (BuildContext context) {
-            switch (routeSettings.name) {
-              case SampleItemDetailsView.routeName:
-                return const SampleItemDetailsView();
-              case SampleItemListView.routeName:
-              default:
-                return const SampleItemListView();
-            }
-          },
-        );
-      },
+      home: const CounterPage(),
     );
   }
 }
